rename pkgcfg
